Restructure to produce "notes" binary from go install
Move entrypoint to cmd/notes/main.go so `go install` produces a binary
named "notes" directly. Move cobra commands to internal/cli/. Rename
module to github.com/dreikanter/notescli to match the repo.
