+++ cmd/notes/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/dreikanter/notescli/internal/cli"
+
+func main() {
+	cli.Execute()
+}
