--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/dreikanter/surat/cmd"
-
-func main() {
-	cmd.Execute()
-}
